back-end/requirements.txt
azure-ai-formrecognizer
fastapi
uvicorn[standard]
python-multipart
requests